requirements.txt
sqlmodel==0.0.8
pytest==7.2.1
SQLAlchemy==1.4.41
fastapi[all]==0.89.1
uvicorn[standard]==0.20.0
gunicorn==20.1.0
alembic==1.9.2
asyncpg==0.27.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytest-asyncio
